Move `fill` and `fill_inplace` to their own module
This brings the number of lines in `lib.rs` down to less than 1k.

The split between unit tests for `fill` and `wrap` could be cleaned
up: right now we test on either level depending on what looks better
in the test.
